# docs/swagger.yaml
basePath: /
definitions:
...
  request.COAGroupCreateRequest:
    properties:
      code:
        minLength: 1
        type: string
      name:
        minLength: 1
        type: string
      normal_balance:
        minLength: 1
        type: string
    required:
    - code
    - name
    - normal_balance
    type: object
  request.COAGroupUpdateRequest:
    properties:
      code:
        maxLength: 20
        minLength: 1
        type: string
      id:
        type: string
      name:
        maxLength: 20
        minLength: 1
        type: string
      normal_balance:
        minLength: 1
        type: string
    required:
    - code
    - name
    - normal_balance
    type: object
  request.COASubGroupCreateRequest:
    properties:
      code:
        minLength: 1
        type: string
      group_id:
        type: string
      name:
        minLength: 1
        type: string
    required:
    - code
    - group_id
    - name
    type: object
  request.COASubGroupUpdateRequest:
    properties:
      code:
        maxLength: 20
        minLength: 1
        type: string
      group_id:
        type: string
      id:
        type: string
      name:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - code
    - group_id
    - name
    type: object
...
  response.JSON:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      status:
        type: integer
    type: object
  response.Links:
    properties:
      current_page:
        type: string
      first_page:
        type: string
      last_page:
        type: string
      next_page:
        type: string
      prev_page:
        type: string
    type: object
  response.Meta:
    properties:
      info:
        type: string
      links:
        $ref: '#/definitions/response.Links'
      page:
        type: integer
      page_size:
        type: integer
      search:
        type: string
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SelectJSON:
    properties:
      data: {}
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: fatihrizqon@gmail.com
    name: Fatih Rizqon
  description: This is an Official Documentation for Go REST API with Fiber Framework
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go REST API with Fiber Framework
  version: "1.0"
paths:
...
  /api/v1/coa_groups:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort column (code, name, normal_balance, status, created_at,
          updated_at)
        in: query
        name: sort
        type: string
      - description: Sort direction (asc, desc)
        in: query
        name: order
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all records.
          schema:
            $ref: '#/definitions/response.JSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Get all chart of account groups
      tags:
      - '#ChartofAccounts Group'
    post:
      consumes:
      - application/json
      parameters:
      - description: COAGroup Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.COAGroupCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: A new record has been stored.
          schema:
            $ref: '#/definitions/response.JSON'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Create chart of account group
      tags:
      - '#ChartofAccounts Group'
  /api/v1/coa_groups/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: COAGroup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Selected record has been deleted.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: COAGroup not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Delete chart of account group
      tags:
      - '#ChartofAccounts Group'
    get:
      consumes:
      - application/json
      parameters:
      - description: COAGroup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved selected record.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: COAGroup not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Get chart of account group by ID
      tags:
      - '#ChartofAccounts Group'
    put:
      consumes:
      - application/json
      parameters:
      - description: COAGroup ID
        in: path
        name: id
        required: true
        type: string
      - description: COAGroup Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.COAGroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Selected record has been updated.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: COAGroup not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Update chart of account group
      tags:
      - '#ChartofAccounts Group'
  /api/v1/coa_subgroups:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort column (code, name, status, created_at, updated_at)
        in: query
        name: sort
        type: string
      - description: Sort direction (asc, desc)
        in: query
        name: order
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all records.
          schema:
            $ref: '#/definitions/response.JSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Get all chart of account subgroups
      tags:
      - '#ChartofAccounts SubGroups'
    post:
      consumes:
      - application/json
      parameters:
      - description: COASubGroup Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.COASubGroupCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: A new record has been stored.
          schema:
            $ref: '#/definitions/response.JSON'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Create chart of account subgroup
      tags:
      - '#ChartofAccounts SubGroups'
  /api/v1/coa_subgroups/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: COASubGroup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Selected record has been deleted.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: COASubGroup not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Delete chart of account subgroup
      tags:
      - '#ChartofAccounts SubGroups'
    get:
      consumes:
      - application/json
      parameters:
      - description: COASubGroup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved selected record.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: COASubGroup not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Get chart of account subgroup by ID
      tags:
      - '#ChartofAccounts SubGroups'
    put:
      consumes:
      - application/json
      parameters:
      - description: COASubGroup ID
        in: path
        name: id
        required: true
        type: string
      - description: COASubGroup Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.COASubGroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Selected record has been updated.
          schema:
            $ref: '#/definitions/response.JSON'
        "404":
          description: COASubGroup not found
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Update chart of account subgroup
      tags:
      - '#ChartofAccounts SubGroups'
...
  /api/v1/dropdown/coa_groups:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved dropdown options
          schema:
            $ref: '#/definitions/response.SelectJSON'
      security:
      - BearerAuth: []
      summary: Get COAGroup dropdown options
      tags:
      - Dropdowns
  /api/v1/dropdown/coa_subgroups:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved dropdown options
          schema:
            $ref: '#/definitions/response.SelectJSON'
      security:
      - BearerAuth: []
      summary: Get COASubGroup dropdown options
      tags:
      - Dropdowns
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"